type module and .js
