+++ package.json
@@ -16,17 +16,18 @@
   "license": "MIT",
   "version": "1.0.7",
   "main": "./dist/index.js",
+  "type": "module",
   "engines": {
     "node": ">=18"
   },
   "scripts": {
     "start": "node dist/index.js",
-    "build": "esbuild src/*.mjs src/**/*.mjs --sourcemap --platform=node --out-extension:.js=.mjs --main-fields=main --outdir=dist --allow-overwrite --analyze",
+    "build": "esbuild src/*.js src/**/*.js --sourcemap --platform=node --main-fields=main --outdir=dist --allow-overwrite --analyze",
     "lint": "eslint ./src",
     "lint:dev": "eslint --fix ./src",
-    "test:base": "node --test --test-force-exit --experimental-test-coverage --test-reporter=spec './src/*.test.mjs' './src/**/*.test.mjs'",
+    "test:base": "node --test --test-force-exit --experimental-test-coverage --test-reporter=spec './src/*.test.js' './src/**/*.test.js'",
     "test": "npm run lint && npm run test:base",
-    "watch:test": "cross-env DEBUG=@natlibfi/* NODE_ENV=test node --watch --test --experimental-test-coverage --test-reporter=spec './src/*.test.mjs' './src/**/*.test.mjs'",
+    "watch:test": "cross-env DEBUG=@natlibfi/* NODE_ENV=test node --watch --test --experimental-test-coverage --test-reporter=spec './src/*.test.js' './src/**/*.test.js'",
     "dev": "npm run watch:test",
     "dev:debug": "cross-env LOG_LEVEL=debug DEBUG=@natlibfi/* NODE_ENV=test npm run watch:test"
   },
@@ -51,9 +52,9 @@
     "uuid": "^11.1.0"
   },
   "devDependencies": {
-    "@natlibfi/fixugen": "^3.0.0-alpha.4",
-    "@natlibfi/fixura": "^4.0.0-alpha.9",
-    "@natlibfi/fixura-mongo": "^3.0.0-alpha.4",
+    "@natlibfi/fixugen": "^3.0.0-alpha.5",
+    "@natlibfi/fixura": "^4.0.0-alpha.12",
+    "@natlibfi/fixura-mongo": "^3.0.0-alpha.5",
     "cross-env": "^7.0.3",
     "esbuild": "0.25.6",
     "eslint": "^9.30.1"
