+++ src/interfaces/index.js
@@ -1,0 +1,2 @@
+export {default as blobsFactory} from './blobs.mjs';
+export {default as profilesFactory} from './profiles.js';
