--- src/interfaces/index.mjs
@@ -1,2 +1,0 @@
-export {default as blobsFactory} from './blobs.mjs';
-export {default as profilesFactory} from './profiles.mjs';
