+++ src/interfaces/profilesQuery.test.js
@@ -5,7 +5,7 @@ import mongoFixturesFactory from '@natlibfi/fixura-mongo';
 import {Error as ApiError} from '@natlibfi/melinda-commons';
 import generateTests from '@natlibfi/fixugen';
 
-import profilesFactory from './profiles.mjs';
+import profilesFactory from './profiles.js';
 
 let mongoFixtures;
