+++ src/interfaces/profilesRemove.test.js
@@ -2,17 +2,16 @@ import {describe} from 'node:test';
 import assert from 'node:assert';
 import {READERS} from '@natlibfi/fixura';
 import mongoFixturesFactory from '@natlibfi/fixura-mongo';
-import {Error as ApiError} from '@natlibfi/melinda-commons';
 import generateTests from '@natlibfi/fixugen';
 
-import profilesFactory from './profiles.mjs';
+import profilesFactory from './profiles.js';
 
 describe('interfaces/profiles', () => {
   let mongoFixtures;
 
   generateTests({
     callback,
-    path: [import.meta.dirname, '..', '..', 'test-fixtures', 'profiles', 'createOrUpdate'],
+    path: [import.meta.dirname, '..', '..', 'test-fixtures', 'profiles', 'remove'],
     recurse: false,
     useMetadataFile: true,
     fixura: {
@@ -37,7 +36,8 @@ describe('interfaces/profiles', () => {
 
   async function initMongofixtures() {
     mongoFixtures = await mongoFixturesFactory({
-      rootPath: [import.meta.dirname, '..', '..', 'test-fixtures', 'profiles', 'createOrUpdate'],
+      rootPath: [import.meta.dirname, '..', '..', 'test-fixtures', 'profiles', 'remove'],
+      gridFS: {bucketName: 'blobmetadatas'},
       useObjectId: true
     });
   }
@@ -45,33 +45,27 @@ describe('interfaces/profiles', () => {
   async function callback({
     getFixture,
     expectToFail = false,
-    expectedFailStatus = '',
-    update = false
+    expectedFailStatus = ''
   }) {
     try {
       const mongoUri = await mongoFixtures.getUri();
-      const createPayload = getFixture('createPayload.json');
-      const user = getFixture('user.json');
+      const dbContents = getFixture('dbContents.json');
       const expectedDb = getFixture('expectedDb.json');
+      const user = getFixture('user.json');
       const profiles = await profilesFactory({MONGO_URI: mongoUri, MONGO_DB: ''});
 
-      await profiles.createOrUpdate({id: 'foo', payload: createPayload, user});
-
-      if (update) {
-        const updatePayload = getFixture('updatePayload.json');
-        await profiles.createOrUpdate({id: 'foo', payload: updatePayload, user});
-      }
+      await mongoFixtures.populate(dbContents);
+      await profiles.remove({id: 'foo', user});
 
       const db = await mongoFixtures.dump();
+
       assert.deepStrictEqual(db.profiles, expectedDb.profiles);
       assert.equal(expectToFail, false, 'This is expected to succes');
     } catch (error) {
       if (!expectToFail) {
         throw error;
       }
-      // console.log(error); // eslint-disable-line
       assert.equal(expectToFail, true, 'This is expected to fail');
-      assert(error instanceof ApiError);
       assert.equal(error.status, expectedFailStatus);
     }
   }
