+++ src/middleware/index.js
@@ -1,0 +1,1 @@
+export {getUserApplicationRoles, generateUserAuthorizationMiddleware, generatePermissionMiddleware} from './authentication.js';
