--- src/middleware/index.mjs
@@ -1,1 +1,0 @@
-export {getUserApplicationRoles, generateUserAuthorizationMiddleware, generatePermissionMiddleware} from './authentication.mjs';
