+++ src/routes/index.js
@@ -1,0 +1,4 @@
+export {default as createBlobsRouter} from './blobs.js';
+export {default as createProfilesRouter} from './profiles.js';
+export {default as createApiDocRouter} from './api-doc.js';
+export {createStatusRouter} from './status.js';
