--- src/routes/index.mjs
@@ -1,4 +1,0 @@
-export {default as createBlobsRouter} from './blobs.mjs';
-export {default as createProfilesRouter} from './profiles.mjs';
-export {default as createApiDocRouter} from './api-doc.mjs';
-export {createStatusRouter} from './status.mjs';
